tests: test_financial_matters_visibility.py
